moved selection_flagged to selection folder
