+++ cpp/oneapi/dal/backend/primitives/selection/BUILD
@@ -33,3 +33,16 @@ dal_test_suite(
         "exclusive",
     ],
 )
+
+dal_test_suite(
+    name = "perf_tests",
+    framework = "catch2",
+    compile_as = [ "dpc++" ],
+    private = True,
+    srcs = glob([
+        "test/*perf_select_flagged*.cpp",
+    ]),
+    dal_deps = [
+        ":selection",
+    ],
+)
