+++ cpp/oneapi/dal/backend/primitives/selection/select_flagged_dpc.cpp
@@ -14,7 +14,7 @@
 * limitations under the License.
 *******************************************************************************/
 
-#include "oneapi/dal/backend/primitives/selection_flagged/select.hpp"
+#include "oneapi/dal/backend/primitives/selection/select_flagged.hpp"
 
 #include <CL/sycl/ONEAPI/experimental/builtins.hpp>
